last commit

+++ src/main/java/com/abonnement/entity/AbonnementSansEngagement.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class AbonnementSansEngagement extends Abonnement {
 
-    public AbonnementSansEngagement(String nomService, int montantMensuel, LocalDate dateDebut, LocalDate dateFin, StatutAbonnement statut) {
+    public AbonnementSansEngagement(String nomService, double montantMensuel, LocalDate dateDebut, LocalDate dateFin, StatutAbonnement statut) {
         super(nomService, montantMensuel, dateDebut, dateFin, statut);
     }
 }
